// src/back/events.enum.ts
export enum Events {
    MESSAGE_CREATED = "message.created",
    MESSAGE_UPDATED = "message.updated",
    MESSAGE_REMOVED = "message.removed",
    REACTION_CREATED = "reaction.created",
    REACTION_UPDATED = "reaction.updated",
    REACTION_REMOVED = "reaction.removed",
    WORKSPACE_MEMBER_ADDED = "workspace.member.added",
    WORKSPACE_MEMBER_REMOVED = "workspace.member.removed",
    WORKSPACE_CREATED = "workspace.created",
    WORKSPACE_UPDATED = "workspace.updated",
    WORKSPACE_REMOVED = "workspace.removed",
    CHANNEL_CREATED = "channel.created",
    CHANNEL_UPDATED = "channel.updated",
    CHANNEL_REMOVED = "channel.removed",
    NOTIFICATION_CREATED = "notification.created",
    NOTIFICATION_READ = "notification.read"
}